Rename swagger package to openapi
This renaming aligns with the ongoing implementation of changes
across the project. By adopting this adjustment, we ensure that
our code remains in line with OpenDaylight's standards.

JIRA: LIGHTY-224
