+++ lighty-applications/lighty-rnc-app-aggregator/lighty-rnc-module/src/main/java/io/lighty/applications/rnc/module/RncLightyModule.java
@@ -26,11 +26,11 @@ import io.lighty.modules.southbound.netconf.impl.NetconfSBPlugin;
 import io.lighty.modules.southbound.netconf.impl.NetconfTopologyPluginBuilder;
 import io.lighty.modules.southbound.netconf.impl.config.NetconfConfiguration;
 import io.lighty.modules.southbound.netconf.impl.util.NetconfConfigUtils;
+import io.lighty.openapi.SwaggerLighty;
 import io.lighty.server.Http2LightyServerBuilder;
 import io.lighty.server.HttpsLightyServerBuilder;
 import io.lighty.server.LightyServerBuilder;
 import io.lighty.server.config.LightyServerConfig;
-import io.lighty.swagger.SwaggerLighty;
 import java.net.InetSocketAddress;
 import java.security.Security;
 import java.util.concurrent.ExecutionException;
