+++ lighty-modules/lighty-openapi/src/main/java/io/lighty/openapi/SwaggerLighty.java
@@ -5,7 +5,7 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at https://www.eclipse.org/legal/epl-v10.html
  */
-package io.lighty.swagger;
+package io.lighty.openapi;
 
 import com.google.common.annotations.VisibleForTesting;
 import io.lighty.core.controller.api.AbstractLightyModule;
